+++ src/components/QuestionDialog/index.ts
@@ -1,2 +1,0 @@
-export { QuestionDialog, type PendingQuestion } from "./QuestionDialog";
-export { default } from "./QuestionDialog";