+++ src/components/Skill/index.ts
@@ -1,3 +1,0 @@
-export { SkillManager } from "./SkillManager";
-export { SkillCard } from "./SkillCard";
-export { SkillSelector } from "./SkillSelector";